Use service class for user syncing
