+++ back/src/main/resources/data-single-test.sql
@@ -36,3 +36,28 @@ INSERT INTO user_role (role_id, user_id) VALUES
 -- DELETE FROM user_role;
 -- DELETE FROM user_historic;
 -- DELETE FROM user_;
+
+-- ===============================
+-- Point states (états des points) STEEVE
+-- ===============================
+INSERT INTO point_state (label, order_, progress) VALUES
+('nouveau', 1.0, 0.0),
+('en cours', 2.0, 0.5),
+('termine', 3.0, 1.0);
+
+-- ===============================
+-- Point types (échelle professionnelle)
+-- ===============================
+INSERT INTO point_type (label) VALUES
+('peu grave'),
+('grave'),
+('tres grave');
+
+INSERT INTO point (date_, surface, budget, coordinates, user_id, point_state_id, point_type_id) VALUES
+(NOW(), null, null, ST_GeomFromText('POINT(47.5260 -18.9095)', 4326), 2, 1, (SELECT id FROM point_type WHERE label='peu grave'));
+
+INSERT INTO factory (label) VALUES
+('Factory A'),
+('Factory B');
+
+INSERT INTO point_historic (date_, surface, budget, coordinates, user_id, point_state_id, point_type_id, point_id) VALUES
